requirements.txt
# These are the main packages needed for the project.
# paho-mqtt is for MQTT communication between sensors and the fog node.
# requests is used for any HTTP calls (e.g. hitting REST endpoints from fog layer).
# flask is the web framework for the fog node's local dashboard/API.
# boto3 is the AWS SDK — needed to talk to DynamoDB, IoT Core, Lambda etc.
# python-dotenv loads environment variables from a .env file so we don't hardcode secrets.

paho-mqtt==1.6.1
requests==2.31.0
flask==3.0.0
boto3==1.34.0
python-dotenv==1.0.0
